profile: set a temporary GOPATH when tentatively running the go tool
This script attempts to run the Go tool in order to detect and
configure a Go installation. It necessarily does this before GOPATH
is set. Since I usually have Go installed at $HOME/go, this conflicts
with the default GOPATH and the go tool issues a warning.

Set GOPATH=/tmp during the Go detection procedure to get rid of the
warning.